+++ packages/client/src/components/auth/sign-up.module.scss
@@ -1,0 +1,189 @@
+@use 'src/styles/buttons';
+
+.signUpFormInput {
+    & > input:focus + label{
+        color:#6910a8 !important;
+    }
+
+    & > input{
+        --box-shadow-color: #6910a818 !important;
+    }
+
+    @media only screen and (max-width: 576px){
+        & > input{
+            width:13rem !important;
+        }
+    }
+}
+
+
+
+.signUpLayout {
+    flex: 1 0 0;
+    display: flex;
+    padding: 2rem 0;
+    flex-direction: column;
+    align-items: center;
+    justify-content:center;
+}
+
+.signUpForm {
+    display:flex;
+    flex-direction:column;
+    align-items: center;
+    width:16rem;
+    padding:0;
+    gap: 0.3rem;
+}
+
+
+.signUpFileUpload {
+    width:17.2rem;
+    position: relative;
+    height: 2rem;
+    margin: 0 0 0.1rem 0;
+    @media only screen and (max-width: 576px){
+        width: 14rem;
+    }
+}
+
+.signUpFileInput {
+    width:100%;
+    opacity:0;
+}
+
+.signUpFileLabel {
+    --box-shadow-color: #6910a818;
+    position:absolute;
+    left:0;
+    top:0;
+    right: 0;
+    height:100%;
+    border:1px solid #0000081d;
+    border-radius: 0.8rem;
+    box-sizing:border-box;
+    display: flex;
+    align-items:center;
+    justify-content: space-between;
+    padding-left: 1rem;
+    line-height:0.8rem;
+
+    &:active{
+        box-shadow: inset 0 0 2px 1px var(--box-shadow-color,#0000081d)
+    }
+
+    &:after{
+        content: 'Upload';
+        height:100%;
+        width: 10%;
+        min-width:6rem;
+        border-radius: 0 0.8rem 0.8rem 0;
+        display:flex;
+        align-items:center;
+        justify-content: center;
+        line-height:0.8rem;
+        background: #4361ee;
+        color: white;
+    }
+
+    &:active:after{
+        box-shadow: 0 0 5px 1px var(--box-shadow-color,#0000081d);
+    }
+}
+
+.signUpFileCancelLabel {
+    position:absolute;
+    
+    bottom: -0.8rem;
+    right: 0.3rem;
+    color: crimson;
+    font-size:0.65rem;
+    cursor: pointer;
+    &:active{
+        text-shadow: 0 0 2px #b1123228;
+    }
+}
+
+.signUpLabel {
+    font-size:0.73rem;
+    align-self: flex-start;
+    margin: -0.3rem 0 0.3rem 0;
+    color:#4361ee;
+    cursor: pointer;
+    &:active{
+        text-shadow: 0 0 1px #4362ee81;
+    }
+    @media only screen and (max-width: 576px){
+        font-size:0.6rem;
+        align-self: center;
+    }
+}
+
+.signUpButton {
+    @extend %button;
+    width: 17rem;
+    border-radius:2rem;
+    margin: 0;
+    background:#4361ee;
+    color: white;
+    border: 1px solid transparent;
+    margin:0.8rem;
+    height: 2.3rem;
+
+    &:hover{
+        border: 1px solid #364fc0;
+        background: #364fc0;
+    }
+
+    &:focus{
+        outline:none;
+        box-shadow: 0 0 8px 2px #4362ee44 ;
+    }
+
+    @media only screen and (max-width: 576px){
+        width: 14rem;
+    }
+}
+
+.signUpPasswordStrengthBar {
+    width:100%;
+    display:flex;
+    align-items:center;
+    margin-top: -0.65rem;
+    max-height:0.35rem;
+    overflow:hidden;
+
+    & > div {
+        
+        flex: 1 0 0;
+        margin: 0;
+    }
+
+    & > p {
+        display:none;
+    }
+
+    @media only screen and (max-width: 576px){
+        width: 13rem;
+    }
+}
+
+.signUpMessageSmallLabel {
+    font-size:0.65rem;
+    align-self: flex-start;
+    color: crimson;
+    margin: -0.35rem 0 0.1rem 0;
+    @media only screen and (max-width: 576px){
+        font-size:0.6rem;
+        align-self: center;
+    }
+}
+
+.signUpMessageLabel {
+    font-size:0.8rem;
+    align-self: center;
+    color: crimson;
+    @media only screen and (max-width: 576px){
+        font-size:0.6rem;
+    }
+}
