+++ packages/client/src/components/post-card.module.scss
@@ -65,22 +65,6 @@
   }
 }
 
-.faPencil {
-  color: black;
-  font-size: 1.7rem;
-  height: 1.2rem;
-  cursor: pointer;
-}
-
-.faTimes {
-  svg {
-    color: crimson;
-    font-size: 1.7rem;
-    height: 2rem;
-    cursor: pointer;
-  }
-}
-
 .toolBar {
   transition: all ease-in 0.3s;
   display: flex;
@@ -88,6 +72,25 @@
   flex-direction: row-reverse;
   gap: 0.3rem;
   opacity: 0;
+
+  &:hover {
+    opacity: 1;
+  }
+}
+
+
+.faPencil {
+  color: black;
+  font-size: 1.7rem;
+  height: 1.2rem !important;
+  cursor: pointer;
+}
+
+.faTimes {
+  color: crimson;
+  font-size: 1.7rem;
+  // height: 2rem ;
+  cursor: pointer;
 }
 
 .postCardHeader {
