+++ packages/client/src/components/styled-component/complex/floating-label-text-input.module.scss
@@ -1,0 +1,46 @@
+@use "src/styles/inputs";
+
+.inputWrapper {
+  position: relative;
+  margin: 0.2rem;
+
+  & > label {
+    display: inline-block;
+    align-self: flex-start;
+    margin: 0.2rem 0;
+    font-weight: 400;
+    font-size: 0.9rem;
+    visibility: hidden;
+    position: absolute;
+    top: 15%;
+    left: 3.7%;
+    transition: all ease-out 0.2s;
+  }
+
+  & > input {
+    @extend %textInput;
+  }
+
+  & > input:focus + label {
+    --text-shadow-color: #ffffff;
+    visibility: visible;
+    display: inline-block;
+    align-self: flex-start;
+    margin: 0.2rem 0;
+    font-size: 0.6rem;
+    position: absolute;
+    top: -25%;
+    left: 3.9%;
+    color: black;
+    text-shadow: 0 0 1px var(--text-shadow-color, white),
+      0 0 2px var(--text-shadow-color, white),
+      0 0 3px var(--text-shadow-color, white),
+      0 0 4px var(--text-shadow-color, white),
+      0 0 5px var(--text-shadow-color, white),
+      0 0 6px var(--text-shadow-color, white),
+      0 0 7px var(--text-shadow-color, white),
+      0 0 8px var(--text-shadow-color, white),
+      0 0 9px var(--text-shadow-color, white),
+      0 0 10px var(--text-shadow-color, white);
+  }
+}
