+++ packages/client/src/components/user-actions.module.scss
@@ -1,0 +1,23 @@
+.userActionCard {
+  position: sticky;
+  top: 1rem;
+  min-width: 14rem;
+  padding: 1rem;
+  border-radius: 1rem;
+  border: 1px solid #6910a8;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 1rem;
+}
+
+.actionItem {
+  text-decoration: none;
+  font-size: 1rem;
+  font-weight: 400;
+  color: #6910a8;
+  outline: none;
+  &:hover {
+    cursor: pointer;
+  }
+}
